Compare multiple navigation objects

Comparison of error analysis plot for different trajectories

compare.navigation(
  ...,
  nsim = 100,
  emu_for_covmat = 1,
  col_50 = "#E74C3C4D",
  col_95 = "#F5B0414D",
  col_50_brd = "#E74C3C",
  col_95_brd = "#F5B041",
  col_traj_error = "#1C12F54D",
  time_interval_simu = 1
)

Arguments

...

navigation objects to be compared.

nsim

An integer indicating the number of trajectories simulated in order to compute the CI

emu_for_covmat

The emulated trajectory for which to use the var-cov matrix in order to simulate data and compute the CI of the error

col_50

The color for the 50% confidence intervals.

col_95

The color for the 95% confidence intervals.

col_50_brd

The color for the 50% confidence intervals borders.

col_95_brd

The color for the 95% confidence intervals.

col_traj_error

The color of the L2 norm of errors of the emulated trajectories.

time_interval_simu

A value in second indicating the interval at which are simulated trajectories in order to compute the CI

Author

Stephane Guerrier, Mehran Khaghani & Lionel Voirol

Examples

if (FALSE) {
#Load trajectory
load("example0.RData")
traj = make_trajectory(data = traj, system = "ned")

# Monte-Carlo settings----------------------------
num.runs1 = 15
num.runs2 = 3

# Timing and sampling frequencies-----------------
timing = make_timing(nav.start     = 0,
                     nav.end       = 50,
                     freq.imu      = 10,
                     freq.gps      = 1,
                     freq.baro     = .5,
                     gps.out.start = 25,
                     gps.out.end   = 45)

# sensor model for data generation----------------
snsr.mdl=list()

imu.freq = 250
acc.mdl = WN(sigma2 = 1.535466e-04) + RW(gamma2 = 1.619511e-10) + DR(omega = 1.276475e-08)
gyr.mdl = WN(sigma2 = 1.711080e-03) + RW(gamma2 = 1.532765e-13)
snsr.mdl$imu = make_sensor(name="imu", frequency=imu.freq, 
error_model1=acc.mdl, error_model2=gyr.mdl)

gps.freq = 1
gps.mdl.pos.hor = WN(sigma2 = 2^2)
gps.mdl.pos.ver = WN(sigma2 = 4^2)
gps.mdl.vel.hor = WN(sigma2 = 0.04^2)
gps.mdl.vel.ver = WN(sigma2 = 0.06^2)
snsr.mdl$gps = make_sensor(name="gps", frequency=gps.freq,
                           error_model1=gps.mdl.pos.hor,
                           error_model2=gps.mdl.pos.ver,
                           error_model3=gps.mdl.vel.hor,
                           error_model4=gps.mdl.vel.ver)
baro.freq = 1
baro.mdl = WN(sigma2=0.5^2)
snsr.mdl$baro = make_sensor(name="baro", frequency=baro.freq, error_model1=baro.mdl)

# sensor model for the KF------------------------
KF.mdl = list()

KF.mdl$imu  = snsr.mdl$imu
KF.mdl$gps  = snsr.mdl$gps
KF.mdl$baro = snsr.mdl$baro

# create navigation object------------------------
x = navigation(traj.ref = traj,
               timing = timing,
               snsr.mdl = snsr.mdl,
               KF.mdl = KF.mdl,
               num.runs = num.runs1)
x2 = navigation(traj.ref = traj,
               timing = timing,
               snsr.mdl = snsr.mdl,
               KF.mdl = KF.mdl,
               num.runs = num.runs2)

# compare navigation objects------------------------
compare.navigation(x1,x2)
}