Construct a trajectory object

Create a trajectory object from simple matrix input.

Usage

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

An object of class trajectory.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n <- 100
dat <- cbind(
  seq(from = 0, to = 60 * 60, length.out = n),
  46.204391 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  6.143158 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  375 + cumsum(rnorm(n))
)
traj <- make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806438177246711 0.107213657930851 374.673668603568
#>   2 36.3636363636364  0.80646332032203 0.107215800889521 374.401962280303
#>   3 72.7272727272727 0.806472425707448 0.107210060124895  376.28091854149
#>   4 109.090909090909 0.806463221361445 0.107207989834317 377.253110714881
#>   5 145.454545454545 0.806459290945937 0.107199816493275 377.437337724386
#> ---                                                                      
#>  96 3454.54545454545 0.806441461286793 0.107270955889347 376.438815477139
#>  97 3490.90909090909 0.806441581204047  0.10727061209588 377.516372180539
#>  98 3527.27272727273 0.806457601929166 0.107265094147239 377.038129314589
#>  99 3563.63636363636 0.806426100648515 0.107276955368873 375.562491837939
#> 100             3600 0.806415551042875 0.107260600233146 376.471902112027
#> 
plot(traj)