Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806407100096149 0.107214833548469 375.720771356135
#>   2 36.3636363636364 0.806420263235267 0.107203487217397 375.206230336698
#>   3 72.7272727272727 0.806403727592061  0.10720835670844 374.998305547862
#>   4 109.090909090909 0.806403299617048 0.107226354390921 374.053722809919
#>   5 145.454545454545 0.806415884613259  0.10722364695661 374.272349871818
#> ---                                                                      
#>  96 3454.54545454545 0.806380096375793 0.107209877176212 381.888805578274
#>  97 3490.90909090909 0.806386983793659  0.10720249972914  382.33480118832
#>  98 3527.27272727273  0.80637149706842 0.107196446383951  384.27706592451
#>  99 3563.63636363636 0.806384469886711 0.107209048287374 382.799998034366
#> 100             3600 0.806388622687975 0.107197114815899 382.846528609541
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806407100096149 0.107214833548469 375.720771356135
#>   2 36.3636363636364 0.806420263235267 0.107203487217397 375.206230336698
#>   3 72.7272727272727 0.806403727592061  0.10720835670844 374.998305547862
#>   4 109.090909090909 0.806403299617048 0.107226354390921 374.053722809919
#>   5 145.454545454545 0.806415884613259  0.10722364695661 374.272349871818
#> ---                                                                      
#>  96 3454.54545454545 0.806380096375793 0.107209877176212 381.888805578274
#>  97 3490.90909090909 0.806386983793659  0.10720249972914  382.33480118832
#>  98 3527.27272727273  0.80637149706842 0.107196446383951  384.27706592451
#>  99 3563.63636363636 0.806384469886711 0.107209048287374 382.799998034366
#> 100             3600 0.806388622687975 0.107197114815899 382.846528609541
#> 
plot(traj)