Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806415590554955 0.107200320154993 374.874253180693
#>   2 36.3636363636364 0.806402644074716 0.107194871210391  375.83235234636
#>   3 72.7272727272727  0.80639343475686 0.107196189980556 374.700332560023
#>   4 109.090909090909 0.806383085005622 0.107203190390531 374.794280088504
#>   5 145.454545454545 0.806394882439061 0.107194736713723 374.623538298339
#> ---                                                                      
#>  96 3454.54545454545 0.806415403677891 0.107187599313127 381.730681393521
#>  97 3490.90909090909 0.806422164482744 0.107204300695482 382.733164210704
#>  98 3527.27272727273  0.80642526731198 0.107223133152369 383.331732128616
#>  99 3563.63636363636 0.806421829865056 0.107229607730753 382.182600248007
#> 100             3600 0.806431092872903 0.107233350374162 381.772358017474
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806415590554955 0.107200320154993 374.874253180693
#>   2 36.3636363636364 0.806402644074716 0.107194871210391  375.83235234636
#>   3 72.7272727272727  0.80639343475686 0.107196189980556 374.700332560023
#>   4 109.090909090909 0.806383085005622 0.107203190390531 374.794280088504
#>   5 145.454545454545 0.806394882439061 0.107194736713723 374.623538298339
#> ---                                                                      
#>  96 3454.54545454545 0.806415403677891 0.107187599313127 381.730681393521
#>  97 3490.90909090909 0.806422164482744 0.107204300695482 382.733164210704
#>  98 3527.27272727273  0.80642526731198 0.107223133152369 383.331732128616
#>  99 3563.63636363636 0.806421829865056 0.107229607730753 382.182600248007
#> 100             3600 0.806431092872903 0.107233350374162 381.772358017474
#> 
plot(traj)