Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80641041089219 0.107199428929866 376.182170694837
#>   2 36.3636363636364 0.806434857425366 0.107202155137361 379.601905906477
#>   3 72.7272727272727  0.80643123816043 0.107218823257786 380.500037525159
#>   4 109.090909090909 0.806426374359658 0.107222139634861 380.998790147591
#>   5 145.454545454545 0.806426111971239 0.107214243498194 381.052031560605
#> ---                                                                      
#>  96 3454.54545454545  0.80657079343103 0.107230560982788 389.791713733837
#>  97 3490.90909090909 0.806562487152586 0.107229264581691 389.533636358722
#>  98 3527.27272727273 0.806563449486888  0.10722932280082 389.123518778697
#>  99 3563.63636363636 0.806570654358111 0.107222061332354 389.321267322289
#> 100             3600 0.806579866153329 0.107225957875488 388.963988532065
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80641041089219 0.107199428929866 376.182170694837
#>   2 36.3636363636364 0.806434857425366 0.107202155137361 379.601905906477
#>   3 72.7272727272727  0.80643123816043 0.107218823257786 380.500037525159
#>   4 109.090909090909 0.806426374359658 0.107222139634861 380.998790147591
#>   5 145.454545454545 0.806426111971239 0.107214243498194 381.052031560605
#> ---                                                                      
#>  96 3454.54545454545  0.80657079343103 0.107230560982788 389.791713733837
#>  97 3490.90909090909 0.806562487152586 0.107229264581691 389.533636358722
#>  98 3527.27272727273 0.806563449486888  0.10722932280082 389.123518778697
#>  99 3563.63636363636 0.806570654358111 0.107222061332354 389.321267322289
#> 100             3600 0.806579866153329 0.107225957875488 388.963988532065
#> 
plot(traj)