Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806424929459101  0.10722359521091 374.825826052352
#>   2 36.3636363636364 0.806436001223421 0.107209748933542 373.736224733968
#>   3 72.7272727272727  0.80644099770451  0.10721212771104 374.510153487174
#>   4 109.090909090909 0.806449461363614 0.107213038764964 374.648467635517
#>   5 145.454545454545 0.806461072237558 0.107198460161926 372.868724827976
#> ---                                                                      
#>  96 3454.54545454545  0.80625961162988 0.107095366389155  374.01969873858
#>  97 3490.90909090909 0.806279559405916  0.10708939409701 372.638285525196
#>  98 3527.27272727273 0.806256092698716   0.1070897415484 371.544067460303
#>  99 3563.63636363636 0.806257615864061 0.107094000347736 371.550315474419
#> 100             3600 0.806274195560266 0.107093576583455 372.119308558206
#> 
plot(traj)