Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80641871026343 0.107202767331249 374.756662769481
#>   2 36.3636363636364 0.806444559137422 0.107208675726431 372.903546996565
#>   3 72.7272727272727  0.80643732569137 0.107206096667358 371.995312129098
#>   4 109.090909090909 0.806452912377321 0.107203853882153 372.207818017289
#>   5 145.454545454545  0.80646108988854  0.10720204562289 373.147598176264
#> ---                                                                      
#>  96 3454.54545454545 0.806585125671769 0.107282813409659 377.578611839714
#>  97 3490.90909090909 0.806576252806951 0.107286022218512 378.898468807788
#>  98 3527.27272727273 0.806558232228326  0.10729292002477 377.867622673737
#>  99 3563.63636363636 0.806584328305343 0.107294084048237 377.960394486328
#> 100             3600 0.806571302703804 0.107294466350093 376.909230771072
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80641871026343 0.107202767331249 374.756662769481
#>   2 36.3636363636364 0.806444559137422 0.107208675726431 372.903546996565
#>   3 72.7272727272727  0.80643732569137 0.107206096667358 371.995312129098
#>   4 109.090909090909 0.806452912377321 0.107203853882153 372.207818017289
#>   5 145.454545454545  0.80646108988854  0.10720204562289 373.147598176264
#> ---                                                                      
#>  96 3454.54545454545 0.806585125671769 0.107282813409659 377.578611839714
#>  97 3490.90909090909 0.806576252806951 0.107286022218512 378.898468807788
#>  98 3527.27272727273 0.806558232228326  0.10729292002477 377.867622673737
#>  99 3563.63636363636 0.806584328305343 0.107294084048237 377.960394486328
#> 100             3600 0.806571302703804 0.107294466350093 376.909230771072
#> 
plot(traj)