Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80643232333876 0.107212568627104 376.623165834441
#>   2 36.3636363636364 0.806439467191886  0.10720799301233 377.014382105173
#>   3 72.7272727272727 0.806427329011693 0.107203496039075 377.818098962456
#>   4 109.090909090909 0.806420151716105 0.107207823793482  376.86607555604
#>   5 145.454545454545 0.806412434720969   0.1072040417573 375.221129346795
#> ---                                                                      
#>  96 3454.54545454545 0.806221994063927 0.107235668952875 376.238175054304
#>  97 3490.90909090909 0.806196277879619 0.107231206804089 376.628773683889
#>  98 3527.27272727273 0.806202504371055 0.107244478329751  376.53193232093
#>  99 3563.63636363636 0.806206396292694 0.107243322976491 376.986520156495
#> 100             3600 0.806206441390084 0.107269741836177 377.048470886242
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")