Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806416369869398 0.107225932636708 373.878794197507
#>   2 36.3636363636364 0.806408706799117 0.107215474935369 374.890104531094
#>   3 72.7272727272727 0.806396137379721 0.107209869433226  374.05350544643
#>   4 109.090909090909 0.806394989503193 0.107217258088707 373.452160831705
#>   5 145.454545454545 0.806384256657364 0.107229181554646 371.741860094187
#> ---                                                                      
#>  96 3454.54545454545 0.806521287391731 0.107302467999599 380.848907696156
#>  97 3490.90909090909 0.806520717065572 0.107294054196096 381.699367858154
#>  98 3527.27272727273 0.806520532829975 0.107281642964908 381.380208563223
#>  99 3563.63636363636 0.806536313514774 0.107283433829573 382.526607668551
#> 100             3600 0.806540653904715 0.107282354982081 384.200262397527
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806416369869398 0.107225932636708 373.878794197507
#>   2 36.3636363636364 0.806408706799117 0.107215474935369 374.890104531094
#>   3 72.7272727272727 0.806396137379721 0.107209869433226  374.05350544643
#>   4 109.090909090909 0.806394989503193 0.107217258088707 373.452160831705
#>   5 145.454545454545 0.806384256657364 0.107229181554646 371.741860094187
#> ---                                                                      
#>  96 3454.54545454545 0.806521287391731 0.107302467999599 380.848907696156
#>  97 3490.90909090909 0.806520717065572 0.107294054196096 381.699367858154
#>  98 3527.27272727273 0.806520532829975 0.107281642964908 381.380208563223
#>  99 3563.63636363636 0.806536313514774 0.107283433829573 382.526607668551
#> 100             3600 0.806540653904715 0.107282354982081 384.200262397527
#> 
plot(traj)