Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80643165589052 0.107226286140424 373.450033019117
#>   2 36.3636363636364  0.80642119190966 0.107219993911087 375.276966895664
#>   3 72.7272727272727  0.80642181399348 0.107218262559483 374.682306020246
#>   4 109.090909090909 0.806432337865958 0.107237605578847 375.377765027037
#>   5 145.454545454545 0.806447579643852 0.107227659093995  374.60944965385
#> ---                                                                      
#>  96 3454.54545454545  0.80652617685386 0.107265712213334 366.330166250051
#>  97 3490.90909090909 0.806525833578048  0.10726965684855 367.621701899869
#>  98 3527.27272727273 0.806515149430239 0.107282171152909 367.537596991752
#>  99 3563.63636363636 0.806517751436382 0.107278073014994 369.064234277756
#> 100             3600 0.806517369506571 0.107282353070058 367.865149809842
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80643165589052 0.107226286140424 373.450033019117
#>   2 36.3636363636364  0.80642119190966 0.107219993911087 375.276966895664
#>   3 72.7272727272727  0.80642181399348 0.107218262559483 374.682306020246
#>   4 109.090909090909 0.806432337865958 0.107237605578847 375.377765027037
#>   5 145.454545454545 0.806447579643852 0.107227659093995  374.60944965385
#> ---                                                                      
#>  96 3454.54545454545  0.80652617685386 0.107265712213334 366.330166250051
#>  97 3490.90909090909 0.806525833578048  0.10726965684855 367.621701899869
#>  98 3527.27272727273 0.806515149430239 0.107282171152909 367.537596991752
#>  99 3563.63636363636 0.806517751436382 0.107278073014994 369.064234277756
#> 100             3600 0.806517369506571 0.107282353070058 367.865149809842
#> 
plot(traj)