Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806418821026771 0.107231128363479 375.352768748217
#>   2 36.3636363636364 0.806410085558665 0.107230634929794 375.292604274136
#>   3 72.7272727272727 0.806399440184859 0.107240669323618 375.142119732971
#>   4 109.090909090909 0.806399929123547 0.107215323488068 376.485379185847
#>   5 145.454545454545 0.806397298152397 0.107216977333027 376.723635093837
#> ---                                                                      
#>  96 3454.54545454545 0.806356440348204 0.107058360593393 386.315682207709
#>  97 3490.90909090909 0.806358695056302 0.107068841522542 384.978523023016
#>  98 3527.27272727273  0.80635268316859 0.107073829238702 386.859481646913
#>  99 3563.63636363636 0.806340776573243 0.107082011662984 387.543946676733
#> 100             3600 0.806347399498496 0.107071511820106 386.650379361532
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806418821026771 0.107231128363479 375.352768748217
#>   2 36.3636363636364 0.806410085558665 0.107230634929794 375.292604274136
#>   3 72.7272727272727 0.806399440184859 0.107240669323618 375.142119732971
#>   4 109.090909090909 0.806399929123547 0.107215323488068 376.485379185847
#>   5 145.454545454545 0.806397298152397 0.107216977333027 376.723635093837
#> ---                                                                      
#>  96 3454.54545454545 0.806356440348204 0.107058360593393 386.315682207709
#>  97 3490.90909090909 0.806358695056302 0.107068841522542 384.978523023016
#>  98 3527.27272727273  0.80635268316859 0.107073829238702 386.859481646913
#>  99 3563.63636363636 0.806340776573243 0.107082011662984 387.543946676733
#> 100             3600 0.806347399498496 0.107071511820106 386.650379361532
#> 
plot(traj)