Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806411851475312 0.107214528345701 375.773662528988
#>   2 36.3636363636364 0.806418016490246 0.107213220018361 375.568481804663
#>   3 72.7272727272727 0.806410789894762 0.107197608746464 374.930422731017
#>   4 109.090909090909 0.806404301091122 0.107204283987968 375.251469108394
#>   5 145.454545454545 0.806414540242729 0.107200054998374 373.506409427637
#> ---                                                                      
#>  96 3454.54545454545 0.806361740250342 0.107251827251841 383.716149834079
#>  97 3490.90909090909   0.8063488826676 0.107281131761951 384.126244648819
#>  98 3527.27272727273 0.806345497414752 0.107282405926863 386.558130944742
#>  99 3563.63636363636 0.806336484350197 0.107272879310232 387.650670342553
#> 100             3600  0.80632678965798 0.107282926162123 388.939949782623
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")