Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806431919901805 0.107235397097507 375.475516751146
#>   2 36.3636363636364 0.806434203450852  0.10722772441606 376.121799141884
#>   3 72.7272727272727 0.806453305873008 0.107224895685266 374.740068086217
#>   4 109.090909090909 0.806470209749343 0.107244861814518 376.591396693847
#>   5 145.454545454545 0.806461449139527 0.107258672318784 376.465139417312
#> ---                                                                      
#>  96 3454.54545454545 0.806404646552807 0.107262767066605 380.417860116008
#>  97 3490.90909090909 0.806400092409014 0.107264969929846 379.148324598536
#>  98 3527.27272727273 0.806408671678484 0.107259825373168 377.029950732453
#>  99 3563.63636363636  0.80640117162681 0.107281636022822 377.344879396519
#> 100             3600 0.806386945257299 0.107273348024785 377.589303110772
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806431919901805 0.107235397097507 375.475516751146
#>   2 36.3636363636364 0.806434203450852  0.10722772441606 376.121799141884
#>   3 72.7272727272727 0.806453305873008 0.107224895685266 374.740068086217
#>   4 109.090909090909 0.806470209749343 0.107244861814518 376.591396693847
#>   5 145.454545454545 0.806461449139527 0.107258672318784 376.465139417312
#> ---                                                                      
#>  96 3454.54545454545 0.806404646552807 0.107262767066605 380.417860116008
#>  97 3490.90909090909 0.806400092409014 0.107264969929846 379.148324598536
#>  98 3527.27272727273 0.806408671678484 0.107259825373168 377.029950732453
#>  99 3563.63636363636  0.80640117162681 0.107281636022822 377.344879396519
#> 100             3600 0.806386945257299 0.107273348024785 377.589303110772
#> 
plot(traj)