Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806430306408909 0.107205282663056 374.530305489937
#>   2 36.3636363636364 0.806453244337878 0.107206598115639 373.775684766595
#>   3 72.7272727272727 0.806448627941053  0.10721038538485 374.675935013856
#>   4 109.090909090909  0.80646370175001  0.10722011533905 375.470758675193
#>   5 145.454545454545 0.806462128849654 0.107226135578687 375.033577887303
#> ---                                                                      
#>  96 3454.54545454545 0.806547415197456 0.107056630719695 385.676290032607
#>  97 3490.90909090909 0.806544783152755 0.107062233431803 386.892827401399
#>  98 3527.27272727273 0.806545951560263 0.107057617048837 387.562842561281
#>  99 3563.63636363636 0.806543669903198 0.107064242725944  387.80667921184
#> 100             3600 0.806554958930723 0.107066854221321 387.505431067047
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806430306408909 0.107205282663056 374.530305489937
#>   2 36.3636363636364 0.806453244337878 0.107206598115639 373.775684766595
#>   3 72.7272727272727 0.806448627941053  0.10721038538485 374.675935013856
#>   4 109.090909090909  0.80646370175001  0.10722011533905 375.470758675193
#>   5 145.454545454545 0.806462128849654 0.107226135578687 375.033577887303
#> ---                                                                      
#>  96 3454.54545454545 0.806547415197456 0.107056630719695 385.676290032607
#>  97 3490.90909090909 0.806544783152755 0.107062233431803 386.892827401399
#>  98 3527.27272727273 0.806545951560263 0.107057617048837 387.562842561281
#>  99 3563.63636363636 0.806543669903198 0.107064242725944  387.80667921184
#> 100             3600 0.806554958930723 0.107066854221321 387.505431067047
#> 
plot(traj)