Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80642848004247 0.107218003269112 375.499676130546
#>   2 36.3636363636364 0.806414979494164 0.107209891917295 375.514515216382
#>   3 72.7272727272727 0.806414713037617 0.107212010781279 373.677388928981
#>   4 109.090909090909 0.806441366562578 0.107220929257007 373.360933335389
#>   5 145.454545454545 0.806445721322547 0.107205363336559  372.44595780818
#> ---                                                                      
#>  96 3454.54545454545 0.806469383841496 0.107262495614649 374.629872616064
#>  97 3490.90909090909 0.806472287435569 0.107246875789906 374.558900716749
#>  98 3527.27272727273 0.806460552840315 0.107252588032113 374.588383054516
#>  99 3563.63636363636 0.806459496431783 0.107252400658701 376.062383652651
#> 100             3600 0.806461683041363 0.107258076859972 374.510710481622
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80642848004247 0.107218003269112 375.499676130546
#>   2 36.3636363636364 0.806414979494164 0.107209891917295 375.514515216382
#>   3 72.7272727272727 0.806414713037617 0.107212010781279 373.677388928981
#>   4 109.090909090909 0.806441366562578 0.107220929257007 373.360933335389
#>   5 145.454545454545 0.806445721322547 0.107205363336559  372.44595780818
#> ---                                                                      
#>  96 3454.54545454545 0.806469383841496 0.107262495614649 374.629872616064
#>  97 3490.90909090909 0.806472287435569 0.107246875789906 374.558900716749
#>  98 3527.27272727273 0.806460552840315 0.107252588032113 374.588383054516
#>  99 3563.63636363636 0.806459496431783 0.107252400658701 376.062383652651
#> 100             3600 0.806461683041363 0.107258076859972 374.510710481622
#> 
plot(traj)