Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806403298114536 0.107198563241369 376.321981027012
#>   2 36.3636363636364 0.806399126442033 0.107204441513386 377.199353996709
#>   3 72.7272727272727 0.806397490323226 0.107215220638316 378.811222292651
#>   4 109.090909090909 0.806399429113121 0.107228530740822 378.505084928169
#>   5 145.454545454545  0.80640309894726 0.107226968268638 378.313233478975
#> ---                                                                      
#>  96 3454.54545454545 0.806304510467401 0.107272287353835  364.03706612351
#>  97 3490.90909090909 0.806307858115374 0.107273502730166 364.230884372645
#>  98 3527.27272727273 0.806299119756855 0.107284418472884 365.287450123027
#>  99 3563.63636363636 0.806300250666802 0.107298656460369  365.66528388258
#> 100             3600 0.806305703722213 0.107286377980985  364.68869608001
#> 
plot(traj)