Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806429203063302   0.1072147781426 375.275319302855
#>   2 36.3636363636364 0.806413926590555 0.107212693400154 374.287698806116
#>   3 72.7272727272727  0.80642386689699 0.107218897220924 375.481356136138
#>   4 109.090909090909 0.806434810117034 0.107228176933668  376.03212318239
#>   5 145.454545454545 0.806442590414416 0.107214807784108 376.100225801895
#> ---                                                                      
#>  96 3454.54545454545 0.806360348100422 0.107308253041839 366.027519248575
#>  97 3490.90909090909  0.80634477205573 0.107321205902139   364.5371592527
#>  98 3527.27272727273 0.806351152184859  0.10730412027558 363.875734973908
#>  99 3563.63636363636   0.8063459311039 0.107309116933707 362.866965758003
#> 100             3600 0.806343438879366 0.107300248746542  362.34646340993
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")