Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80643131732448 0.107243806420595 375.465370912897
#>   2 36.3636363636364 0.806430874744445 0.107251962021883 375.247619213779
#>   3 72.7272727272727 0.806421209843544 0.107244676913774 376.196189219178
#>   4 109.090909090909 0.806429187564444  0.10724018448157 377.410465101969
#>   5 145.454545454545 0.806430194641787 0.107234254656152 379.064043221452
#> ---                                                                      
#>  96 3454.54545454545 0.806415444182892 0.107011316663393 384.785595966062
#>  97 3490.90909090909 0.806407532981399 0.107031631816102  385.93597622059
#>  98 3527.27272727273 0.806397127436624 0.107042475889248 386.226135250181
#>  99 3563.63636363636 0.806389661333631 0.107060867452886 388.594673197504
#> 100             3600 0.806384670056635 0.107057172021802 389.132141197101
#> 
plot(traj)