Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80640744163155 0.107216105337146 373.969209297316
#>   2 36.3636363636364 0.806419935920984 0.107218313721701 373.573804087973
#>   3 72.7272727272727 0.806409798578436 0.107200092522431 372.155019710483
#>   4 109.090909090909 0.806408619379151 0.107200041841786 373.002011597739
#>   5 145.454545454545 0.806408773867353 0.107208077036544 373.651014255345
#> ---                                                                      
#>  96 3454.54545454545 0.806496301678189 0.107059501296304 375.023849062755
#>  97 3490.90909090909 0.806496323392042 0.107058717396161 374.420403844048
#>  98 3527.27272727273 0.806494083329872 0.107044415124024 375.103436011162
#>  99 3563.63636363636 0.806490691853875 0.107047763683533 374.714510655799
#> 100             3600 0.806505515598973  0.10703196652784 373.527341658572
#> 
plot(traj)