Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419066030642 0.107210834192929  375.14497448321
#>   2 36.3636363636364 0.806407602897865 0.107203963413134 375.983360318191
#>   3 72.7272727272727 0.806398986048088 0.107209297330865 375.850519918359
#>   4 109.090909090909  0.80640703031037 0.107202130634674 374.588113276537
#>   5 145.454545454545 0.806399394944057 0.107190869168405 376.061515169672
#> ---                                                                      
#>  96 3454.54545454545 0.806402317676721 0.107165487464264 372.748593952472
#>  97 3490.90909090909 0.806388708831366 0.107169940232569 372.021732013838
#>  98 3527.27272727273 0.806389312411224  0.10716777625265 372.534769101383
#>  99 3563.63636363636 0.806398410826174 0.107174685676565 372.858508346176
#> 100             3600  0.80640306979273 0.107173029596083 372.317421406426
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419066030642 0.107210834192929  375.14497448321
#>   2 36.3636363636364 0.806407602897865 0.107203963413134 375.983360318191
#>   3 72.7272727272727 0.806398986048088 0.107209297330865 375.850519918359
#>   4 109.090909090909  0.80640703031037 0.107202130634674 374.588113276537
#>   5 145.454545454545 0.806399394944057 0.107190869168405 376.061515169672
#> ---                                                                      
#>  96 3454.54545454545 0.806402317676721 0.107165487464264 372.748593952472
#>  97 3490.90909090909 0.806388708831366 0.107169940232569 372.021732013838
#>  98 3527.27272727273 0.806389312411224  0.10716777625265 372.534769101383
#>  99 3563.63636363636 0.806398410826174 0.107174685676565 372.858508346176
#> 100             3600  0.80640306979273 0.107173029596083 372.317421406426
#> 
plot(traj)