Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806431683942703 0.107206051816908 373.888361399777
#>   2 36.3636363636364 0.806433877287604  0.10721097694455 373.623941784403
#>   3 72.7272727272727  0.80643475457635 0.107203130946344 372.401208032358
#>   4 109.090909090909 0.806427765863433  0.10721208258885 371.172655402419
#>   5 145.454545454545 0.806447625845312 0.107195234932832 371.850980737237
#> ---                                                                      
#>  96 3454.54545454545 0.806577106849439 0.107190433582064 379.535630894437
#>  97 3490.90909090909 0.806584461689883 0.107183457222132 380.732870013649
#>  98 3527.27272727273 0.806576306584833 0.107191448211171 381.453835261484
#>  99 3563.63636363636 0.806572273332378 0.107198153425888 381.269117616835
#> 100             3600 0.806555349218188 0.107196306035028 381.786314956664
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806431683942703 0.107206051816908 373.888361399777
#>   2 36.3636363636364 0.806433877287604  0.10721097694455 373.623941784403
#>   3 72.7272727272727  0.80643475457635 0.107203130946344 372.401208032358
#>   4 109.090909090909 0.806427765863433  0.10721208258885 371.172655402419
#>   5 145.454545454545 0.806447625845312 0.107195234932832 371.850980737237
#> ---                                                                      
#>  96 3454.54545454545 0.806577106849439 0.107190433582064 379.535630894437
#>  97 3490.90909090909 0.806584461689883 0.107183457222132 380.732870013649
#>  98 3527.27272727273 0.806576306584833 0.107191448211171 381.453835261484
#>  99 3563.63636363636 0.806572273332378 0.107198153425888 381.269117616835
#> 100             3600 0.806555349218188 0.107196306035028 381.786314956664
#> 
plot(traj)