Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806427912003938 0.107217019704652 373.992003386249
#>   2 36.3636363636364 0.806433212141735 0.107212946959022 373.381224379226
#>   3 72.7272727272727 0.806414094332267 0.107220477437142 373.870086446117
#>   4 109.090909090909 0.806412109476497 0.107201496561265 372.296097323455
#>   5 145.454545454545 0.806412402667637 0.107209874660743  372.35944685339
#> ---                                                                      
#>  96 3454.54545454545 0.806367581734974 0.107233558410992 372.553810631538
#>  97 3490.90909090909 0.806392480337801  0.10723751807391 372.590564093267
#>  98 3527.27272727273 0.806392711620937  0.10724947722187 373.313242362563
#>  99 3563.63636363636 0.806388761517077 0.107261345471778 373.912529217274
#> 100             3600 0.806384983132228 0.107262523022367 374.488641347681
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806427912003938 0.107217019704652 373.992003386249
#>   2 36.3636363636364 0.806433212141735 0.107212946959022 373.381224379226
#>   3 72.7272727272727 0.806414094332267 0.107220477437142 373.870086446117
#>   4 109.090909090909 0.806412109476497 0.107201496561265 372.296097323455
#>   5 145.454545454545 0.806412402667637 0.107209874660743  372.35944685339
#> ---                                                                      
#>  96 3454.54545454545 0.806367581734974 0.107233558410992 372.553810631538
#>  97 3490.90909090909 0.806392480337801  0.10723751807391 372.590564093267
#>  98 3527.27272727273 0.806392711620937  0.10724947722187 373.313242362563
#>  99 3563.63636363636 0.806388761517077 0.107261345471778 373.912529217274
#> 100             3600 0.806384983132228 0.107262523022367 374.488641347681
#> 
plot(traj)