Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80640313054645 0.107221981023594 374.378789362591
#>   2 36.3636363636364 0.806409964574615 0.107234819387015 374.703160300665
#>   3 72.7272727272727 0.806428386950901 0.107251163285566 373.300153658632
#>   4 109.090909090909 0.806426705230894 0.107242526711724 373.100549868838
#>   5 145.454545454545 0.806442744759339 0.107251070732909 372.830937844193
#> ---                                                                      
#>  96 3454.54545454545 0.806463680176889 0.107313516682882 378.816335734939
#>  97 3490.90909090909 0.806464903116611 0.107311637839372 377.779558085115
#>  98 3527.27272727273 0.806452952586339 0.107291965214392 376.218618341316
#>  99 3563.63636363636 0.806436502387896 0.107274415070344 374.625390515636
#> 100             3600 0.806431052970984 0.107282351100464 374.602372641956
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80640313054645 0.107221981023594 374.378789362591
#>   2 36.3636363636364 0.806409964574615 0.107234819387015 374.703160300665
#>   3 72.7272727272727 0.806428386950901 0.107251163285566 373.300153658632
#>   4 109.090909090909 0.806426705230894 0.107242526711724 373.100549868838
#>   5 145.454545454545 0.806442744759339 0.107251070732909 372.830937844193
#> ---                                                                      
#>  96 3454.54545454545 0.806463680176889 0.107313516682882 378.816335734939
#>  97 3490.90909090909 0.806464903116611 0.107311637839372 377.779558085115
#>  98 3527.27272727273 0.806452952586339 0.107291965214392 376.218618341316
#>  99 3563.63636363636 0.806436502387896 0.107274415070344 374.625390515636
#> 100             3600 0.806431052970984 0.107282351100464 374.602372641956
#> 
plot(traj)