Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806415981565452 0.107211737018916 376.291129506514
#>   2 36.3636363636364 0.806414285320965 0.107202394267497 376.075060163186
#>   3 72.7272727272727  0.80640058809554 0.107203871960379 376.658479838775
#>   4 109.090909090909 0.806409389890738 0.107201641192246 377.281720367116
#>   5 145.454545454545 0.806401858097447 0.107203469863557 378.078645563066
#> ---                                                                      
#>  96 3454.54545454545 0.806532167948142 0.107226732510295 376.132223762122
#>  97 3490.90909090909 0.806525447019681 0.107226890373418 375.371809078368
#>  98 3527.27272727273 0.806546833480388 0.107238805487558 374.652135459838
#>  99 3563.63636363636 0.806567933974623 0.107233749680465 373.530271140235
#> 100             3600 0.806579621307964 0.107244600593619 373.103233032586
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806415981565452 0.107211737018916 376.291129506514
#>   2 36.3636363636364 0.806414285320965 0.107202394267497 376.075060163186
#>   3 72.7272727272727  0.80640058809554 0.107203871960379 376.658479838775
#>   4 109.090909090909 0.806409389890738 0.107201641192246 377.281720367116
#>   5 145.454545454545 0.806401858097447 0.107203469863557 378.078645563066
#> ---                                                                      
#>  96 3454.54545454545 0.806532167948142 0.107226732510295 376.132223762122
#>  97 3490.90909090909 0.806525447019681 0.107226890373418 375.371809078368
#>  98 3527.27272727273 0.806546833480388 0.107238805487558 374.652135459838
#>  99 3563.63636363636 0.806567933974623 0.107233749680465 373.530271140235
#> 100             3600 0.806579621307964 0.107244600593619 373.103233032586
#> 
plot(traj)