Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806420611332806 0.107212268235599 376.316358926412
#>   2 36.3636363636364 0.806413403220643   0.1072072829206 376.306800555303
#>   3 72.7272727272727 0.806420790648998  0.10720524042973 375.380961794609
#>   4 109.090909090909  0.80642940299966 0.107199958594017 377.099973433276
#>   5 145.454545454545 0.806436763766871 0.107195782977217 377.074951448641
#> ---                                                                      
#>  96 3454.54545454545 0.806637748672813 0.107156932990015 382.584386817426
#>  97 3490.90909090909 0.806622393443058 0.107164313643645 383.246106198866
#>  98 3527.27272727273 0.806600527120081 0.107157490129806 383.340036344322
#>  99 3563.63636363636 0.806603997505188 0.107166708505304 384.133196977785
#> 100             3600 0.806613143500674 0.107169159772535 383.066802739984
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806420611332806 0.107212268235599 376.316358926412
#>   2 36.3636363636364 0.806413403220643   0.1072072829206 376.306800555303
#>   3 72.7272727272727 0.806420790648998  0.10720524042973 375.380961794609
#>   4 109.090909090909  0.80642940299966 0.107199958594017 377.099973433276
#>   5 145.454545454545 0.806436763766871 0.107195782977217 377.074951448641
#> ---                                                                      
#>  96 3454.54545454545 0.806637748672813 0.107156932990015 382.584386817426
#>  97 3490.90909090909 0.806622393443058 0.107164313643645 383.246106198866
#>  98 3527.27272727273 0.806600527120081 0.107157490129806 383.340036344322
#>  99 3563.63636363636 0.806603997505188 0.107166708505304 384.133196977785
#> 100             3600 0.806613143500674 0.107169159772535 383.066802739984
#> 
plot(traj)