Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806430295617046 0.107228659769484 375.157472724358
#>   2 36.3636363636364 0.806426320736222  0.10722273735733 374.933740445994
#>   3 72.7272727272727 0.806427108513707  0.10721520476931 375.940655788505
#>   4 109.090909090909  0.80644365167449 0.107235732053326 376.124312674789
#>   5 145.454545454545 0.806459069452103 0.107238168123361 376.947084151982
#> ---                                                                      
#>  96 3454.54545454545 0.806357956316021 0.107144351316891 386.747624482011
#>  97 3490.90909090909 0.806345556491854 0.107137894112706 388.033127919237
#>  98 3527.27272727273 0.806336507819917 0.107135082544462 387.934108088331
#>  99 3563.63636363636  0.80631968967184 0.107142589278165 389.663589532385
#> 100             3600 0.806310451048613 0.107130778555317 389.658734424512
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806430295617046 0.107228659769484 375.157472724358
#>   2 36.3636363636364 0.806426320736222  0.10722273735733 374.933740445994
#>   3 72.7272727272727 0.806427108513707  0.10721520476931 375.940655788505
#>   4 109.090909090909  0.80644365167449 0.107235732053326 376.124312674789
#>   5 145.454545454545 0.806459069452103 0.107238168123361 376.947084151982
#> ---                                                                      
#>  96 3454.54545454545 0.806357956316021 0.107144351316891 386.747624482011
#>  97 3490.90909090909 0.806345556491854 0.107137894112706 388.033127919237
#>  98 3527.27272727273 0.806336507819917 0.107135082544462 387.934108088331
#>  99 3563.63636363636  0.80631968967184 0.107142589278165 389.663589532385
#> 100             3600 0.806310451048613 0.107130778555317 389.658734424512
#> 
plot(traj)