Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806436588066022 0.107208462752824 375.478076269223
#>   2 36.3636363636364 0.806423224120336 0.107217945409217 377.355337648284
#>   3 72.7272727272727 0.806425820196492 0.107210360319016 378.778950236671
#>   4 109.090909090909 0.806417751031278 0.107212057236165 377.286848111657
#>   5 145.454545454545 0.806405465694834 0.107189902114429 377.697407852055
#> ---                                                                      
#>  96 3454.54545454545 0.806384797331705 0.107294589898498  384.84746111645
#>  97 3490.90909090909 0.806367882456046 0.107292504918783 384.651983498281
#>  98 3527.27272727273 0.806376557712395 0.107288647375056 385.623257195169
#>  99 3563.63636363636 0.806384293172712 0.107290296394576  387.27263882118
#> 100             3600 0.806391868258725 0.107299307252441 389.378852304079
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806436588066022 0.107208462752824 375.478076269223
#>   2 36.3636363636364 0.806423224120336 0.107217945409217 377.355337648284
#>   3 72.7272727272727 0.806425820196492 0.107210360319016 378.778950236671
#>   4 109.090909090909 0.806417751031278 0.107212057236165 377.286848111657
#>   5 145.454545454545 0.806405465694834 0.107189902114429 377.697407852055
#> ---                                                                      
#>  96 3454.54545454545 0.806384797331705 0.107294589898498  384.84746111645
#>  97 3490.90909090909 0.806367882456046 0.107292504918783 384.651983498281
#>  98 3527.27272727273 0.806376557712395 0.107288647375056 385.623257195169
#>  99 3563.63636363636 0.806384293172712 0.107290296394576  387.27263882118
#> 100             3600 0.806391868258725 0.107299307252441 389.378852304079
#> 
plot(traj)