Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806414321825019  0.10721952172926 373.931720207958
#>   2 36.3636363636364 0.806413346315254 0.107205702605473 372.548072717101
#>   3 72.7272727272727 0.806412014450801  0.10720213460168 373.998567531209
#>   4 109.090909090909 0.806419450217681  0.10719588137413 373.831879473405
#>   5 145.454545454545 0.806413395096193 0.107169575742363 376.637823011309
#> ---                                                                      
#>  96 3454.54545454545 0.806381952974533 0.107132422210828 389.088552891911
#>  97 3490.90909090909 0.806381558786808 0.107112128561362  388.77979673937
#>  98 3527.27272727273 0.806382628622839 0.107110995681848 390.123137144699
#>  99 3563.63636363636 0.806404054643538 0.107109615389324 391.154026218418
#> 100             3600 0.806418278474767 0.107095805434409 390.234818366647
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806414321825019  0.10721952172926 373.931720207958
#>   2 36.3636363636364 0.806413346315254 0.107205702605473 372.548072717101
#>   3 72.7272727272727 0.806412014450801  0.10720213460168 373.998567531209
#>   4 109.090909090909 0.806419450217681  0.10719588137413 373.831879473405
#>   5 145.454545454545 0.806413395096193 0.107169575742363 376.637823011309
#> ---                                                                      
#>  96 3454.54545454545 0.806381952974533 0.107132422210828 389.088552891911
#>  97 3490.90909090909 0.806381558786808 0.107112128561362  388.77979673937
#>  98 3527.27272727273 0.806382628622839 0.107110995681848 390.123137144699
#>  99 3563.63636363636 0.806404054643538 0.107109615389324 391.154026218418
#> 100             3600 0.806418278474767 0.107095805434409 390.234818366647
#> 
plot(traj)