Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n <- 100
dat <- cbind(
  seq(from = 0, to = 60 * 60, length.out = n),
  46.204391 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  6.143158 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  375 + cumsum(rnorm(n))
)
traj <- make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419890288627 0.107210706141035 375.839192952704
#>   2 36.3636363636364 0.806421296515993 0.107227471964286  375.26373994261
#>   3 72.7272727272727  0.80643024125095 0.107243234662551 374.988199862924
#>   4 109.090909090909 0.806454873060177 0.107259230471074  377.81936681822
#>   5 145.454545454545  0.80646789760883 0.107252113009118 375.736842959958
#> ---                                                                      
#>  96 3454.54545454545  0.80649843772094 0.107288637328075 363.162705471342
#>  97 3490.90909090909 0.806501015042028 0.107277007606937 363.503779647189
#>  98 3527.27272727273 0.806517398778496 0.107274181045419 365.574685401089
#>  99 3563.63636363636 0.806520442204167 0.107280521167619  362.96972348189
#> 100             3600 0.806526596953379 0.107285895975029 363.710572959028
#> 
plot(traj)