Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806408112753966 0.107218148855096 373.334134506743
#>   2 36.3636363636364 0.806415628876383 0.107209147696965 371.467257754056
#>   3 72.7272727272727 0.806424400113817 0.107209932817541 371.505636500369
#>   4 109.090909090909 0.806422988147066 0.107207775886642 371.874031793486
#>   5 145.454545454545 0.806418427675977 0.107189918359237 370.210686477655
#> ---                                                                      
#>  96 3454.54545454545 0.806506205397674 0.107296568247024 362.684016235771
#>  97 3490.90909090909 0.806505053506269 0.107292452789424 361.449864203833
#>  98 3527.27272727273 0.806508348590111 0.107290900398451 359.670864797191
#>  99 3563.63636363636 0.806512658388897 0.107283065930536 359.767548346472
#> 100             3600 0.806514785073543 0.107286565948344  360.83709103789
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806408112753966 0.107218148855096 373.334134506743
#>   2 36.3636363636364 0.806415628876383 0.107209147696965 371.467257754056
#>   3 72.7272727272727 0.806424400113817 0.107209932817541 371.505636500369
#>   4 109.090909090909 0.806422988147066 0.107207775886642 371.874031793486
#>   5 145.454545454545 0.806418427675977 0.107189918359237 370.210686477655
#> ---                                                                      
#>  96 3454.54545454545 0.806506205397674 0.107296568247024 362.684016235771
#>  97 3490.90909090909 0.806505053506269 0.107292452789424 361.449864203833
#>  98 3527.27272727273 0.806508348590111 0.107290900398451 359.670864797191
#>  99 3563.63636363636 0.806512658388897 0.107283065930536 359.767548346472
#> 100             3600 0.806514785073543 0.107286565948344  360.83709103789
#> 
plot(traj)