Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806432921913423 0.107202180199887 373.910510230817
#>   2 36.3636363636364 0.806439822955203 0.107193198572363   373.8262702445
#>   3 72.7272727272727 0.806433307806254 0.107206151495083 373.410343494133
#>   4 109.090909090909 0.806431864519455 0.107212345234575 373.418668061719
#>   5 145.454545454545  0.80644261264555 0.107224872129906 372.963436283278
#> ---                                                                      
#>  96 3454.54545454545 0.806349187976996  0.10739939350531 379.197686170323
#>  97 3490.90909090909 0.806343490903617 0.107400901898939 379.161995800762
#>  98 3527.27272727273 0.806333893964189 0.107393430490192  378.85270064541
#>  99 3563.63636363636 0.806341168876598 0.107394725119954 378.728191888443
#> 100             3600 0.806343405296449 0.107397149267442 378.437127565226
#> 
plot(traj)