Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0  0.80641480043429 0.107202806154444 374.123291872488
#>   2 36.3636363636364 0.806424746852819 0.107218691309984 372.348992475793
#>   3 72.7272727272727 0.806437012415116 0.107218974686376 371.582408437967
#>   4 109.090909090909 0.806432336979656 0.107210246809959 372.126100137403
#>   5 145.454545454545 0.806414167942801 0.107205400056799 372.870603069174
#> ---                                                                      
#>  96 3454.54545454545 0.806265253815252 0.107298194449598 357.322025604662
#>  97 3490.90909090909  0.80624988295918 0.107287756919515 357.086777134706
#>  98 3527.27272727273 0.806257946275386 0.107282475532571  357.33305616042
#>  99 3563.63636363636  0.80627121518807 0.107295609773054 358.005063324175
#> 100             3600 0.806278245482533 0.107294304606757 358.412978917858
#> 
plot(traj)