Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806427677224705 0.107212145754855 374.779082900798
#>   2 36.3636363636364 0.806416290339497 0.107212384645045 376.633206864681
#>   3 72.7272727272727  0.80641073721769 0.107224219101473 376.232087528656
#>   4 109.090909090909 0.806415349468651  0.10723837705231 376.560389443435
#>   5 145.454545454545 0.806417140252496 0.107229077335398 375.448273426355
#> ---                                                                      
#>  96 3454.54545454545  0.80660379200524 0.107331099658341 372.616863659742
#>  97 3490.90909090909 0.806602575560409 0.107337827545287 372.002670328661
#>  98 3527.27272727273 0.806594865022567 0.107339963680896 371.994946285532
#>  99 3563.63636363636 0.806584770607386 0.107335969363367 373.424315535101
#> 100             3600 0.806595331437159 0.107338765426304 373.443991404547
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806427677224705 0.107212145754855 374.779082900798
#>   2 36.3636363636364 0.806416290339497 0.107212384645045 376.633206864681
#>   3 72.7272727272727  0.80641073721769 0.107224219101473 376.232087528656
#>   4 109.090909090909 0.806415349468651  0.10723837705231 376.560389443435
#>   5 145.454545454545 0.806417140252496 0.107229077335398 375.448273426355
#> ---                                                                      
#>  96 3454.54545454545  0.80660379200524 0.107331099658341 372.616863659742
#>  97 3490.90909090909 0.806602575560409 0.107337827545287 372.002670328661
#>  98 3527.27272727273 0.806594865022567 0.107339963680896 371.994946285532
#>  99 3563.63636363636 0.806584770607386 0.107335969363367 373.424315535101
#> 100             3600 0.806595331437159 0.107338765426304 373.443991404547
#> 
plot(traj)