Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806427835890636 0.107220976604094 375.573636732709
#>   2 36.3636363636364 0.806442385791516 0.107235401822248 375.722064682714
#>   3 72.7272727272727 0.806457338366105 0.107220479199234 375.454073219389
#>   4 109.090909090909 0.806458671792755 0.107213273935304  376.39389697831
#>   5 145.454545454545 0.806457488098278 0.107226429016845 375.334929967659
#> ---                                                                      
#>  96 3454.54545454545 0.806489755427059 0.107202657272467 378.338679752962
#>  97 3490.90909090909 0.806481030982995 0.107207081617722 378.896437772867
#>  98 3527.27272727273 0.806482590479701 0.107209126045984 379.166981456987
#>  99 3563.63636363636 0.806476917169542 0.107211113390881 378.463342235013
#> 100             3600 0.806470950320279 0.107219904597847 378.477338803854
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806427835890636 0.107220976604094 375.573636732709
#>   2 36.3636363636364 0.806442385791516 0.107235401822248 375.722064682714
#>   3 72.7272727272727 0.806457338366105 0.107220479199234 375.454073219389
#>   4 109.090909090909 0.806458671792755 0.107213273935304  376.39389697831
#>   5 145.454545454545 0.806457488098278 0.107226429016845 375.334929967659
#> ---                                                                      
#>  96 3454.54545454545 0.806489755427059 0.107202657272467 378.338679752962
#>  97 3490.90909090909 0.806481030982995 0.107207081617722 378.896437772867
#>  98 3527.27272727273 0.806482590479701 0.107209126045984 379.166981456987
#>  99 3563.63636363636 0.806476917169542 0.107211113390881 378.463342235013
#> 100             3600 0.806470950320279 0.107219904597847 378.477338803854
#> 
plot(traj)