Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806421182401876 0.107217374696886 376.097918181548
#>   2 36.3636363636364 0.806430909746834 0.107210356249059 377.618698719078
#>   3 72.7272727272727 0.806435315702907 0.107204487115862 378.372003653545
#>   4 109.090909090909 0.806434380239709 0.107201481921547 379.273854638086
#>   5 145.454545454545  0.80644043462409 0.107208789834122 379.665031985148
#> ---                                                                      
#>  96 3454.54545454545 0.806446782397013 0.107311803735487    403.077226414
#>  97 3490.90909090909 0.806438244252799 0.107305015133532 402.899608215248
#>  98 3527.27272727273 0.806424934946079 0.107317118597233 402.432544747119
#>  99 3563.63636363636 0.806425990261854 0.107326480774944 402.809988259163
#> 100             3600 0.806433323033447 0.107314781502156 400.934744881703
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")