Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n <- 100
dat <- cbind(
  seq(from = 0, to = 60 * 60, length.out = n),
  46.204391 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  6.143158 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  375 + cumsum(rnorm(n))
)
traj <- make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806402852419875 0.107223083512867 376.631439852667
#>   2 36.3636363636364 0.806412000283863 0.107219111342339 377.354356810101
#>   3 72.7272727272727 0.806400421502282 0.107234518947886 376.091565221133
#>   4 109.090909090909 0.806404762132788 0.107225199073781 375.539582200043
#>   5 145.454545454545 0.806422676257548 0.107245013897535 374.286458105913
#> ---                                                                      
#>  96 3454.54545454545 0.806711130928983 0.107271954905353 367.534855776468
#>  97 3490.90909090909 0.806704731498532  0.10726743284873  367.82932777149
#>  98 3527.27272727273 0.806688973468483 0.107269062982403 369.403619242734
#>  99 3563.63636363636 0.806685797042057  0.10727286141672 368.474163021594
#> 100             3600 0.806681422882361 0.107270790808741 367.948597684295
#> 
plot(traj)