Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419105562138 0.107220876563564  375.16533111497
#>   2 36.3636363636364 0.806428954413571   0.1072331008596 375.639559946572
#>   3 72.7272727272727 0.806410212100351 0.107223300544844 375.402427036583
#>   4 109.090909090909 0.806403578196669 0.107232157709727 375.966041531239
#>   5 145.454545454545 0.806408453093002 0.107224105788406  374.67869710069
#> ---                                                                      
#>  96 3454.54545454545 0.806449110272499 0.107259166066433 366.687245665238
#>  97 3490.90909090909  0.80644995657944 0.107260921275066 368.566379278549
#>  98 3527.27272727273 0.806438642005532 0.107279168174525 369.484606138605
#>  99 3563.63636363636 0.806439334610425 0.107274391775656 370.251389522012
#> 100             3600 0.806431064164361  0.10726308490738 370.574375250369
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419105562138 0.107220876563564  375.16533111497
#>   2 36.3636363636364 0.806428954413571   0.1072331008596 375.639559946572
#>   3 72.7272727272727 0.806410212100351 0.107223300544844 375.402427036583
#>   4 109.090909090909 0.806403578196669 0.107232157709727 375.966041531239
#>   5 145.454545454545 0.806408453093002 0.107224105788406  374.67869710069
#> ---                                                                      
#>  96 3454.54545454545 0.806449110272499 0.107259166066433 366.687245665238
#>  97 3490.90909090909  0.80644995657944 0.107260921275066 368.566379278549
#>  98 3527.27272727273 0.806438642005532 0.107279168174525 369.484606138605
#>  99 3563.63636363636 0.806439334610425 0.107274391775656 370.251389522012
#> 100             3600 0.806431064164361  0.10726308490738 370.574375250369
#> 
plot(traj)