Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419694142301 0.107212068817244 375.454657659838
#>   2 36.3636363636364  0.80641053283596 0.107209905394609 376.274418351182
#>   3 72.7272727272727 0.806416619842345 0.107219032345195 375.067772247737
#>   4 109.090909090909 0.806411476444276 0.107212805137739 375.328372568144
#>   5 145.454545454545 0.806415427255205 0.107234405467916 374.333249178268
#> ---                                                                      
#>  96 3454.54545454545 0.806435664068418 0.107130425203368 378.987268844782
#>  97 3490.90909090909 0.806445099063838  0.10714254977087 380.186915670331
#>  98 3527.27272727273  0.80642845296906 0.107144615129131 380.846229561888
#>  99 3563.63636363636 0.806423569273445 0.107140250853799  381.59896475755
#> 100             3600 0.806429316165003 0.107138443812131 382.768116623897
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806419694142301 0.107212068817244 375.454657659838
#>   2 36.3636363636364  0.80641053283596 0.107209905394609 376.274418351182
#>   3 72.7272727272727 0.806416619842345 0.107219032345195 375.067772247737
#>   4 109.090909090909 0.806411476444276 0.107212805137739 375.328372568144
#>   5 145.454545454545 0.806415427255205 0.107234405467916 374.333249178268
#> ---                                                                      
#>  96 3454.54545454545 0.806435664068418 0.107130425203368 378.987268844782
#>  97 3490.90909090909 0.806445099063838  0.10714254977087 380.186915670331
#>  98 3527.27272727273  0.80642845296906 0.107144615129131 380.846229561888
#>  99 3563.63636363636 0.806423569273445 0.107140250853799  381.59896475755
#> 100             3600 0.806429316165003 0.107138443812131 382.768116623897
#> 
plot(traj)