Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806402292230051 0.107206719662711  374.69674590575
#>   2 36.3636363636364 0.806394645507873 0.107196844449477 373.732409531706
#>   3 72.7272727272727  0.80640881227101 0.107198728669702 373.670858003996
#>   4 109.090909090909 0.806422828936401 0.107193135853346 372.124417808618
#>   5 145.454545454545 0.806418145463924  0.10720163262029 373.477529787453
#> ---                                                                      
#>  96 3454.54545454545 0.806273147993298 0.107155816826627 378.590136151025
#>  97 3490.90909090909 0.806253020387427 0.107163320412179 378.055013082906
#>  98 3527.27272727273 0.806244334601928  0.10715537452829 380.713061548303
#>  99 3563.63636363636 0.806248203498398 0.107163235347696 382.362724619175
#> 100             3600 0.806256935484697 0.107152743378209 381.049800477474
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806402292230051 0.107206719662711  374.69674590575
#>   2 36.3636363636364 0.806394645507873 0.107196844449477 373.732409531706
#>   3 72.7272727272727  0.80640881227101 0.107198728669702 373.670858003996
#>   4 109.090909090909 0.806422828936401 0.107193135853346 372.124417808618
#>   5 145.454545454545 0.806418145463924  0.10720163262029 373.477529787453
#> ---                                                                      
#>  96 3454.54545454545 0.806273147993298 0.107155816826627 378.590136151025
#>  97 3490.90909090909 0.806253020387427 0.107163320412179 378.055013082906
#>  98 3527.27272727273 0.806244334601928  0.10715537452829 380.713061548303
#>  99 3563.63636363636 0.806248203498398 0.107163235347696 382.362724619175
#> 100             3600 0.806256935484697 0.107152743378209 381.049800477474
#> 
plot(traj)