Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806416288921621 0.107223076482399 374.515385196078
#>   2 36.3636363636364 0.806428261493974  0.10722345884552 374.686976805775
#>   3 72.7272727272727 0.806422071090156 0.107224586288782 375.230274132359
#>   4 109.090909090909 0.806410115617912 0.107220893474304 373.630426950918
#>   5 145.454545454545 0.806404200050761 0.107216419844834 374.862491299733
#> ---                                                                      
#>  96 3454.54545454545  0.80643378540171 0.107140924868688 360.194462960144
#>  97 3490.90909090909 0.806417615323757  0.10713487230059  359.44454486273
#>  98 3527.27272727273 0.806416997044399 0.107122875928556  360.42221436806
#>  99 3563.63636363636 0.806408245194534 0.107117034496395 361.446414138822
#> 100             3600 0.806414826632132 0.107107172987763 360.479334036971
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806416288921621 0.107223076482399 374.515385196078
#>   2 36.3636363636364 0.806428261493974  0.10722345884552 374.686976805775
#>   3 72.7272727272727 0.806422071090156 0.107224586288782 375.230274132359
#>   4 109.090909090909 0.806410115617912 0.107220893474304 373.630426950918
#>   5 145.454545454545 0.806404200050761 0.107216419844834 374.862491299733
#> ---                                                                      
#>  96 3454.54545454545  0.80643378540171 0.107140924868688 360.194462960144
#>  97 3490.90909090909 0.806417615323757  0.10713487230059  359.44454486273
#>  98 3527.27272727273 0.806416997044399 0.107122875928556  360.42221436806
#>  99 3563.63636363636 0.806408245194534 0.107117034496395 361.446414138822
#> 100             3600 0.806414826632132 0.107107172987763 360.479334036971
#> 
plot(traj)