Construct trajectory object

Create trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Stephane Guerrier, Mehran Khaghani, and Lionel Voirol

Examples

n = 100
dat = cbind(seq(from = 0, to = 60*60, length.out = n), 
46.204391*pi/180 + cumsum(rnorm(n))/10^5, 
6.143158*pi/180 + cumsum(rnorm(n))/10^5, 
375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806415426143956 0.107223768115526 375.608180431568
#>   2 36.3636363636364 0.806407132476442 0.107214612674918 376.580318534129
#>   3 72.7272727272727 0.806396578642604  0.10723161561505 376.201287178649
#>   4 109.090909090909 0.806394101062417 0.107216363162982 374.835362473832
#>   5 145.454545454545 0.806377251285018 0.107211047354881 374.518929175654
#> ---                                                                      
#>  96 3454.54545454545 0.806315546395687 0.107330610514631 367.246334090169
#>  97 3490.90909090909 0.806313731297549 0.107336045058184 366.396502573129
#>  98 3527.27272727273 0.806312435837201 0.107349278219438 366.116035656013
#>  99 3563.63636363636 0.806309327051335 0.107358837571568 365.465931492964
#> 100             3600 0.806303503283625 0.107353024141136 365.994133178113
#> 
plot(traj)

traj = make_trajectory(data = dat, name = "My cool data", start_time = "13:15:00")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806415426143956 0.107223768115526 375.608180431568
#>   2 36.3636363636364 0.806407132476442 0.107214612674918 376.580318534129
#>   3 72.7272727272727 0.806396578642604  0.10723161561505 376.201287178649
#>   4 109.090909090909 0.806394101062417 0.107216363162982 374.835362473832
#>   5 145.454545454545 0.806377251285018 0.107211047354881 374.518929175654
#> ---                                                                      
#>  96 3454.54545454545 0.806315546395687 0.107330610514631 367.246334090169
#>  97 3490.90909090909 0.806313731297549 0.107336045058184 366.396502573129
#>  98 3527.27272727273 0.806312435837201 0.107349278219438 366.116035656013
#>  99 3563.63636363636 0.806309327051335 0.107358837571568 365.465931492964
#> 100             3600 0.806303503283625 0.107353024141136 365.994133178113
#> 
plot(traj)