Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n <- 100
dat <- cbind(
  seq(from = 0, to = 60 * 60, length.out = n),
  46.204391 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  6.143158 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  375 + cumsum(rnorm(n))
)
traj <- make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806418319109001 0.107222462796353 373.439967344986
#>   2 36.3636363636364 0.806419210911991 0.107216706814029 373.639731797682
#>   3 72.7272727272727 0.806420239532068 0.107206104258184 373.232031506981
#>   4 109.090909090909 0.806404514343334 0.107222490098735 371.538761866525
#>   5 145.454545454545 0.806403467924827 0.107213485377357  372.44639183897
#> ---                                                                      
#>  96 3454.54545454545   0.8063344125172 0.107177983870528 369.048140675362
#>  97 3490.90909090909 0.806318081752978 0.107178283238399 369.067244096774
#>  98 3527.27272727273 0.806307743931168 0.107171984598191 369.310086549098
#>  99 3563.63636363636 0.806300871341816 0.107172856647142 369.722559619317
#> 100             3600 0.806294105023211 0.107169281941626 371.350820246903
#> 
plot(traj)