Construct a trajectory object

Create a trajectory object from simple matrix input.

make_trajectory(
  data,
  system = "ellipsoidal",
  start_time = NULL,
  name = NULL,
  ...
)

Arguments

data

A multiple-column matrix. The first column corresponds to the measurment time (in seconds); columns 2, 3 and 4 corresponds to the positions (with the order lat, long and alt (in rad) if ellipsoidal coord or x_N, x_E and x_D for NED coord); columns 5, 6 and 7 (optional) corresponds to the attitude (with the order roll, pitch and yaw); columns 8, 9 and 10 (optional) corresponds to the velocity along the same axes are columns 2, 3 and 4.

system

A string corresponding to the coordinate system (possible choices: ellipsoidal or ned) considered.

start_time

A string (optional) corresponding to the start time for the trajectory.

name

A string (optional) corresponding to the name of the dataset.

...

Additional arguments.

Value

A trajectory object.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier

Examples

n <- 100
dat <- cbind(
  seq(from = 0, to = 60 * 60, length.out = n),
  46.204391 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  6.143158 * pi / 180 + cumsum(rnorm(n)) / 10^5,
  375 + cumsum(rnorm(n))
)
traj <- make_trajectory(data = dat, name = "My cool data")
traj
#> Data Name: My cool data 
#> 
#> Data preview:
#> 
#>                 time               lat               lon              alt
#>   1                0 0.806425124888075  0.10722150482719 373.551803505782
#>   2 36.3636363636364 0.806404244191091 0.107212729995349  374.16131795017
#>   3 72.7272727272727 0.806409506767064 0.107212799426263  375.64779072619
#>   4 109.090909090909 0.806397611390622 0.107208747545904 375.598697919764
#>   5 145.454545454545 0.806398933913353 0.107210386881456  375.70072993702
#> ---                                                                      
#>  96 3454.54545454545 0.806402583343142 0.107102819359451 382.714887213379
#>  97 3490.90909090909  0.80641650719415 0.107107208990378 383.561604919184
#>  98 3527.27272727273 0.806427410374292 0.107102299330961 383.643672076424
#>  99 3563.63636363636 0.806416818469406 0.107095765998879 383.822598168752
#> 100             3600 0.806414347395261 0.107099221138555 384.521350169571
#> 
plot(traj)