Runs "IMU model evaluation" or "INS-GPS-Baro integrated navigation (sensor fusion)"

This function performs of the two following main tasks, base on the provided input. If a reference trajectory (traj.ref) is provided, it generates sensor data (IMU, GPS, Baro) corrupted by additive errors according to snsr.mdl, and performs navigation using KF.mdl as the sensor error model within the Kalman filter to evaluate how this particular model performs when navigating.

navigation(
  traj.ref,
  timing,
  snsr.mdl,
  KF.mdl,
  g = 9.8056,
  num.runs = 1,
  results.system = "ned",
  x_o = NULL,
  noProgressBar = FALSE,
  IC = NULL,
  imu_data = NULL,
  gps_data = NULL,
  baro_data = NULL,
  input.seed = 0,
  PhiQ_method = "exact",
  P_subsampling = 1,
  compute_PhiQ_each_n = 1,
  parallel.ncores = detectCores(all.tests = FALSE, logical = TRUE),
  tmpdir = tempdir()
)

Arguments

traj.ref

A trajectory object (see the documentation for make_trajectory), serving as the reference trajectory for generating sensor data and evaluating the error in navigation once performed. Only position and attitude data are required/considered, and velocity will be calculated from position.

timing

A timing object (see the documentation for make_timing) containing timing information such as start and end of navigation.

snsr.mdl

A sensor object (see the documentation for make_sensor) containing additive sensor error model to generate realistic sensor data.

KF.mdl

A sensor object (see the documentation for make_sensor) containing additive sensor error model to be used within the Kalman filter for navigation.

g

Gravitational acceleration.

num.runs

Number of times the sensor data generation and navigation is performed (Monte-Carlo simulation).

results.system

The coordinate system (ned/ellipsoidal) in which the results are reported (see the documentation for make_trajectory).

x_o

Origin of the fixed ned frame.

noProgressBar

A bolean specifying if there should not be a progress bar

IC

Initial conditions. See the examples for the format.

imu_data

IMU data. See the examples for the format.

gps_data

GPS data. See the examples for the format.

baro_data

Baro data. See the examples for the format.

input.seed

Seed for the random number generator. Actual seed is computed as input.seed * num.runs + run

PhiQ_method

String that specify the method to compute Phi and Q matrices, can be "exact" or the order of the Taylor expansions to use

P_subsampling

(memory optimization) store only one sample of the P matrix each P_subsampling time instants

compute_PhiQ_each_n

Specify the interval of IMU measurements between each computation of PhiQ.

parallel.ncores

The number of cores to be used for parallel Monte-Carlo runs

tmpdir

Where to store temporary navigation output. It should not be mapped on a filesystem which lives in RAM.

Value

An object of navigation class containing the reference trajectory, fused trajectory, sensor data, covariance matrix, and time.

Author

Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier